Address CodeRabbit review
